 find-house-near-you/requirements.txt
telethon>=1.28.5
openai>=0.27.0
geopy>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
selenium>=4.14.0
webdriver-manager>=3.8.6
